More snippety snip.
Added .snip format documentation to snip-mode description.

Added edit-snippets to easily load up the snippet buffer for a particular
mode+config scope combo. If said file doesn't exist, it is populated with some
tips on how to get started.

Removed the stock snippets from the search path. The expected use case now is
for you to edit your own global snippets for a mode and `insert-stock-snippets`
therein, so that you can both peruse them to see what's available, and tweak
the formatting to your preference.